Move environment.go to test dir
