+++ services/horizon/internal/test/environment.go
@@ -1,4 +1,4 @@
-package horizon
+package test
 
 import (
 	"fmt"
